Normalizes values to fairly compare against tuning benchmarks
When tuning a target system, multiple benchmarks with distinct
magnitudes where given the same weight in the calculation of the
accuracy. This commit changes things so that the speedup w.r.t. the base
system is used instead. This has the implication that magnitude
differences between different benchmarks are removed.

Also, the speedup function is reflected when speedup < 1.0. This is with
the goal of making the MonotonicSearch heuristic work.

Fixes #8