Modifies tunning logic of porta
Now that we've found that cpu-quota/cpu-period affects the way in which
memguard behaves, we need to take this into account in the tuning logic
of porta. The changes in docker/porta reflect this. Basically, we tune
CPU first, then apply those CPU parameters to the benchmarks that are
used to tune memory.

This commit also creates a 'tune-targets' playbook that is in charge of
this task. We then create a 'base-vs-limited-target' playbook that
executes benchmarks on a target with/without limits.