Parametrizes the tuning process
Now the docker image used for tunning can be provided via a flag, as
well as the flags used when invoking 'docker run'. Also, 'max-cpu-quota'
and 'max-mem-bw' control the upper limits that porta uses while tunning
the targets. The helper script is simplified and assumes it's only used
for tuning.

Also updates the tune-targets playbook accordingly