Adds same-platform variability results on issdm cluster with 4 cores
This runs stress-ng on all 4 cores for 30 seconds and reports results

We see that variability is NOT reduced by limiting cpu bandwidth. We
give half of CPU time for each core. This might be due to the way in
which cgroups does multi-core CPU bandwidth limiting. We need to
double-check if this is actually what happens