Results of running base-vs-targets for stressng on 4 machines
Base machine is 'issdm-12' and targets were tunned with the 'crafty' and
'c-ray' benchmarks.

The main difference between these results and the ones that appear on
our VarSys '16 paper is that we are reflecting the speedup function for
x=1, for both (1) tunning targets and (2) displaying results. This
allows us to show better the reduction in variability without having to
deal with different scales (slowdowns that lie between the [0-1] range
are instead reflected and treated as speedups).

In short, we now unambiguously observe reduced variability when targets
are limited. A couple of outliers, in particular stress-ng's memfd
stressor, when limited, is very slow on target machines.